+++ src/mantle/http-client/class-factory.php
@@ -20,6 +20,13 @@ class Factory {
 	}
 
 	/**
+	 * Create a new pending request.
+	 */
+	public static function create(): Pending_Request {
+		return ( new static() )->new_pending_request();
+	}
+
+	/**
 	 * Generate a new pending request.
 	 */
 	protected function new_pending_request(): Pending_Request {