+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `classname`/`the_classname` helper to generate complex class names.
 
+## Fixed
+
+- Fixed issue with command jobs not working properly.
+
 ## v1.0.7 - 2024-04-29
 
 ### Added