+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v1.1.0 - ____-__-__
+## v1.1.0
+
+- Added a `classname`/`the_classname` helper to generate complex class names.
+- Added support for installing the Redis `object-cache.php` drop-in during
+  testing with `with_object_cache()`.
 
 ### Changed
 
@@ -16,6 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   configuration when booting the application. This allows for more slimmer
 	configuration files in the application. Provider service providers will always
 	load without needing to be declared in the application configuration.
+
+## Fixed
+
+- Fixed issue with command jobs not working properly.
+- Ensure that unit tests fail when a project's installation script fails.
+- Fix anonymous queue jobs from WP-CLI failing to run.
 
 ## v1.0.7 - 2024-04-29